# Remove dependance on global window and document