@dead-claudia 2024-07-22
I've got a better idea:

1. Replace `$doc` with `elem.parentDocument` everywhere and just drop the otherwise unused `$window`.
2. Remove the closure `mithril/render/render` currently uses and just export the returned render function directly.

This comes with some added benefits:

1. No need for `$doc.defaultView`
2. You can render inside iframe windows from outside them and it'll select the right window to render to.