@KoryNunn 2024-07-22
Yep sounds great, I'd suggest doing it in stages as changing the export is a breaking change for anyone who uses render directly ( like me :) )

IMO: 
1. Get this in as it's completely non-breaking, and a step in the right direction/intent.
2. PR/Test using parentDocument, I'm not familiar enough with the codebase to be certain that it's non-breaking
3. Propose/PR unwrapping render.