@KoryNunn 2024-07-22
I had a look at only using parent.ownerDocument but that's not viable since DocumentFragment doesn't have an owner document.

Best case we could use parent.ownerDocument when it's available and fall back to the top-level render target for when parent is a DocumentFragment.

This (and my PR) does still leave an issue where render would no longer be able to target DocumentFragments. Maybe a solution to this is to keep the code as is, but use the target node if no document is passed in.