@KoryNunn 2024-07-22
I've updated to implement the last solution I proposed.

This adds a number of fn calls that may have a slight per impact, we could inline if required.