@dead-claudia 2024-07-22
> I had a look at only using parent.ownerDocument but that's not viable since DocumentFragment doesn't have an owner document.

Any browser that returns `null` for `ownerDocument` document fragments is violating the spec: https://dom.spec.whatwg.org/#dom-node-ownerdocument

And this has been the case since DOM Level 1, back in 1998: https://www.w3.org/TR/1998/REC-DOM-Level-1-19981001/level-one-core.html#ID-1950641247

[DOM Level 2 created one other exception, `DocumentType` nodes](https://www.w3.org/TR/2000/REC-DOM-Level-2-Core-20001113/core.html#node-ownerDoc), but those have since been removed entirely in the living standard.