@dead-claudia 2024-07-22
@KoryNunn Forgot to ping. Also, WPT tests for the `new DocumentFragment()` constructor variant: https://github.com/web-platform-tests/wpt/blob/0a9f334462/dom/nodes/DocumentFragment-constructor.html#L13

It's difficult to imagine how `new DocumentFragment()` would differ from `document.createDocumentFragment()`. (And to be honest, it's surprising this isn't tested for directly.)

You can safely it exists for all nodes with a `nodeType` other than `Node.DOCUMENT_NODE === 9` or `Node.DOCUMENT_TYPE_NODE === 10`. And of these, documents forward element insertions to their `documentElement` and doctypes don't accept children at all.

If you truly want something bulletproof, you can use this code at the top of the render function and then proceed to use the current node's `ownerDocument` in place of everywhere:

```js
// The condition is equivalent to the following, but saves some space and is slightly faster:
// `root.nodeType !== 1 && root.nodeType !== 10`
if (~0x402 >> root.nodeType & 1) {
    // Set the real root to an element if a document node
    if (root.nodeType === 9) {
        root = root.documentElement
    } else {
        throw new TypeError("Root must be a document, an element, or a document fragment")
    }
}
```