@KoryNunn 2024-07-22
Yep I was pretty surprised that DocumentFragment didn't have ownerDocument, and it turns out I just forgot to add it to the mockDom for DocumentFragment

PR now only uses parent.ownerDocument
All tests passing.

Closure still exists as there are other variables that seem necessary .