@dead-claudia 2024-07-26
Also, can you run the benchmark suite before and after this patch and share the results here? I expect this to have minor impact on create but to be insignificant elsewhere, and I'd like to have some validation here.