@KoryNunn 2024-07-29
Updated.

Perf results:

Before:

```
> mithril@2.2.3 perf
> node performance/test-perf.js

construct large vnode tree x 41,669 ops/sec ±0.81% (131 runs sampled)
rerender identical vnode x 11,902,057 ops/sec ±2.00% (122 runs sampled)
rerender same tree x 97,702 ops/sec ±1.11% (127 runs sampled)
add large nested tree x 12,689 ops/sec ±1.06% (124 runs sampled)
mutate styles/properties x 151 ops/sec ±1.60% (105 runs sampled)
repeated add/removal x 5,487 ops/sec ±2.00% (123 runs sampled)
Completed perf tests in 49874ms
```

After:
```
> mithril@2.2.2 perf
> node performance/test-perf.js

construct large vnode tree x 40,003 ops/sec ±1.66% (128 runs sampled)
rerender identical vnode x 11,305,752 ops/sec ±0.95% (131 runs sampled)
rerender same tree x 101,118 ops/sec ±1.40% (116 runs sampled)
add large nested tree x 13,239 ops/sec ±1.09% (126 runs sampled)
mutate styles/properties x 150 ops/sec ±2.36% (104 runs sampled)
repeated add/removal x 5,820 ops/sec ±2.19% (118 runs sampled)
Completed perf tests in 49948ms
```